server/requirements.txt
beautifulsoup4
Flask
Flask-Cors
openai
python-dotenv
requests
ruff
